pyproject.toml
[project]
name = "codestory"
version = "0.1.0"
description = "Transform code repositories into audio narratives"
readme = "README.md"
authors = [
    { name = "Nick", email = "krzemienski@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    # Claude Agent SDK (includes anthropic client)
    "claude-agent-sdk>=0.1.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    # HTTP client
    "httpx>=0.28.0",
    # Validation & Config
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9.0",  # TOTP for 2FA
    # Task queue
    "celery[redis]>=5.4.0",
    # AWS S3
    "boto3>=1.35.0",
    # Supabase
    "supabase>=2.10.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codestory"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "UP047", "B010"]  # UP047/B010: TypeVar patterns for mypy compat

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
